reorg windows files
