+++ src/internal/windows/chrome_windows.js
@@ -10,7 +10,7 @@ const which = require("which")
 
 const REASON_CHROME_NOT_DETECTED = `Chrome not detected`
 
-const getCertificateTrustInfoFromChrome = ({ logger, windowsTrustInfo }) => {
+export const executeTrustQueryOnChrome = ({ logger, windowsTrustInfo }) => {
   const chromeDetected = detectChrome({ logger })
   if (!chromeDetected) {
     return {
@@ -23,46 +23,6 @@ const getCertificateTrustInfoFromChrome = ({ logger, windowsTrustInfo }) => {
     status: windowsTrustInfo.status,
     reason: windowsTrustInfo.reason,
   }
-}
-
-const addCertificateInChromeTrustStore = ({ logger, windowsTrustInfo, existingTrustInfo }) => {
-  if (existingTrustInfo && existingTrustInfo.chrome.status === "other") {
-    return existingTrustInfo.chrome
-  }
-
-  const chromeDetected = detectChrome({ logger })
-  if (!chromeDetected) {
-    return {
-      status: "other",
-      reason: REASON_CHROME_NOT_DETECTED,
-    }
-  }
-
-  return {
-    status: windowsTrustInfo.status,
-    reason: windowsTrustInfo.reason,
-  }
-}
-
-const removeCertificateFromChromeTrustStore = ({ logger, windowsTrustInfo }) => {
-  const chromeDetected = detectChrome({ logger })
-  if (!chromeDetected) {
-    return {
-      status: "other",
-      reason: REASON_CHROME_NOT_DETECTED,
-    }
-  }
-
-  return {
-    status: windowsTrustInfo.status,
-    reason: windowsTrustInfo.reason,
-  }
-}
-
-export const chromeTrustStoreOnWindows = {
-  getCertificateTrustInfo: getCertificateTrustInfoFromChrome,
-  addCertificate: addCertificateInChromeTrustStore,
-  removeCertificate: removeCertificateFromChromeTrustStore,
 }
 
 // https://github.com/litixsoft/karma-detect-browsers/blob/332b4bdb2ab3db7c6a1a6d3ec5a1c6ccf2332c4d/browsers/Chrome.js#L1
