+++ src/internal/windows/edge.js
@@ -1,0 +1,6 @@
+export const executeTrustQueryOnEdge = ({ windowsTrustInfo }) => {
+  return {
+    status: windowsTrustInfo.status,
+    reason: windowsTrustInfo.reason,
+  }
+}
